requirements.txt
django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
drf-yasg==1.21.7
factory_boy==3.3.0
